bio/fragments: add Iterator.Read, deprecate iterator.Next
Summary:
Add bio.fragments.Iterator.Read, which returns the next fragment in a newly allocated  memory. If the caller wants to recycle the memory, it should call fragments.PutInFreePool.

This API is the same as that for BAM/PAM, and I believe it's easier to use, less error prone. It should be also slightly faster because we can remove a typedmemmove call.

I'll convert existing uses of Next() incrementally.

Reviewers: joshnewman

Reviewed By: joshnewman

Differential Revision: https://phabricator.grailbio.com/D29461

fbshipit-source-id: 7a10ebf